Делаит проект на модули
